Split PixelClusterTagInfoProducer in .cc and .h files
