+++ RecoBTag/PixelCluster/plugins/PixelClusterTagInfoProducer.cc
@@ -1,113 +1,6 @@
-// -*- C++ -*-
-//
-// Package:    RecoBTag/PixelCluster
-// Class:      PixelClusterTagInfoProducer
-// 
-/**\class PixelClusterTagInfoProducer PixelCluster RecoBTag/PixelCluster/plugins/PixelClusterTagInfoProducer.cc
 
- Description: Produces a vector of PixelClusterTagInfo objects.
+#include "RecoBTag/PixelCluster/interface/PixelClusterTagInfoProducer.h"
 
- Implementation:
-     [Notes on implementation]
-*/
-//
-// Original Author:  Alberto Zucchetta (UniZ) [zucchett]
-//         Created:  Wed, 03 Jul 2019 12:37:30 GMT
-//
-//
-
-
-// system include files
-#include <memory>
-
-// user include files
-#include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/stream/EDProducer.h"
-
-#include "FWCore/Framework/interface/Event.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-#include "FWCore/Utilities/interface/StreamID.h"
-
-#include "FWCore/Framework/interface/ESHandle.h"
-#include "FWCore/Framework/interface/EventSetup.h"
-
-// TagInfo
-#include "DataFormats/BTauReco/interface/PixelClusterTagInfo.h"
-
-// For vertices
-#include "DataFormats/VertexReco/interface/Vertex.h"
-#include "DataFormats/VertexReco/interface/VertexFwd.h"
-
-// For jet
-#include "DataFormats/JetReco/interface/PFJet.h" 
-#include "DataFormats/JetReco/interface/PFJetCollection.h"
-
-// For pixel clusters
-#include "DataFormats/SiPixelCluster/interface/SiPixelCluster.h"
-#include "DataFormats/Common/interface/DetSetVectorNew.h"
-#include "DataFormats/Common/interface/Ref.h"
-#include "DataFormats/DetId/interface/DetId.h"
-#include "DataFormats/TrackerCommon/interface/TrackerTopology.h"
-
-// Pixel topology
-#include "DataFormats/SiPixelDetId/interface/PixelSubdetector.h"
-#include "DataFormats/SiPixelDetId/interface/PixelBarrelName.h"
-#include "DataFormats/SiPixelDetId/interface/PixelEndcapName.h"
-#include "DataFormats/SiPixelDetId/interface/PXBDetId.h"
-#include "DataFormats/SiPixelDetId/interface/PXFDetId.h"
-
-// Geometry
-#include "Geometry/Records/interface/IdealGeometryRecord.h"
-#include "Geometry/TrackerGeometryBuilder/interface/PixelGeomDetUnit.h"
-#include "Geometry/TrackerGeometryBuilder/interface/PixelGeomDetType.h"
-#include "Geometry/TrackerGeometryBuilder/interface/TrackerGeometry.h"
-#include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
-#include "Geometry/CommonDetUnit/interface/GeomDetType.h"
-#include "Geometry/CommonTopologies/interface/PixelTopology.h"
-
-#include "TVector3.h"
-#include "TLorentzVector.h"
-
-
-//
-// class declaration
-//
-
-class PixelClusterTagInfoProducer : public edm::stream::EDProducer<> {
-   public:
-      explicit PixelClusterTagInfoProducer(const edm::ParameterSet&);
-      ~PixelClusterTagInfoProducer();
-
-      static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
-
-   private:
-
-      virtual void beginStream(edm::StreamID) override;
-      virtual void produce(edm::Event&, const edm::EventSetup&) override;
-      virtual void endStream() override;
-
-      //virtual void beginRun(edm::Run const&, edm::EventSetup const&) override;
-      //virtual void endRun(edm::Run const&, edm::EventSetup const&) override;
-      //virtual void beginLuminosityBlock(edm::LuminosityBlock const&, edm::EventSetup const&) override;
-      //virtual void endLuminosityBlock(edm::LuminosityBlock const&, edm::EventSetup const&) override;
-      
-//      virtual void fillData(reco::PixelClusterData*, float, float, int, int);
-
-      // ----------member data ---------------------------
-      edm::ParameterSet iConfig;
-      
-      edm::EDGetTokenT<edm::View<reco::Jet> >                 m_jets;
-      edm::EDGetTokenT<reco::VertexCollection >               m_vertices;
-      edm::EDGetTokenT<edmNew::DetSetVector<SiPixelCluster> > m_pixelhit;
-      bool m_isPhase1;
-      bool m_addFPIX;
-      int m_minADC;
-      double m_minJetPt;
-      double m_maxJetEta;
-      int m_nLayers;
-};
 
 //
 // constants, enums and typedefs
@@ -405,7 +298,7 @@ PixelClusterTagInfoProducer::produce(edm::Event& iEvent, const edm::EventSetup& 
 //    
 //}
 
-
+/*
 // ------------ method called once each stream before processing any runs, lumis or events  ------------
 void
 PixelClusterTagInfoProducer::beginStream(edm::StreamID)
@@ -416,7 +309,7 @@ PixelClusterTagInfoProducer::beginStream(edm::StreamID)
 void
 PixelClusterTagInfoProducer::endStream() {
 }
-
+*/
 // ------------ method called when starting to processes a run  ------------
 /*
 void
@@ -448,7 +341,7 @@ PixelClusterTagInfoProducer::endLuminosityBlock(edm::LuminosityBlock const&, edm
 {
 }
 */
- 
+/*
 // ------------ method fills 'descriptions' with the allowed parameters for the module  ------------
 void
 PixelClusterTagInfoProducer::fillDescriptions(edm::ConfigurationDescriptions& descriptions) {
@@ -458,6 +351,7 @@ PixelClusterTagInfoProducer::fillDescriptions(edm::ConfigurationDescriptions& de
   desc.setUnknown();
   descriptions.addDefault(desc);
 }
-
+*/
 //define this as a plug-in
 DEFINE_FWK_MODULE(PixelClusterTagInfoProducer);
+
